// ui/models/generated/createUser.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Schema for creating a user
 */
export interface CreateUser {
  /**
   * The name of the user
   */
  name: string;
  /**
   * The email of the user
   */
  email: string;
  /**
   * The type of authentication
   */
  authType: "basic" | "google";
  /**
   * The mechanism of authentication
   */
  authMechanism?: Basic | Google;
}
/**
 * Basic authentication schema
 */
export interface Basic {
  /**
   * The type of authentication
   */
  type?: "basic";
  /**
   * The password of the user
   */
  password: string;
  /**
   * The password of the user
   */
  confirmPassword: string;
}
/**
 * Basic authentication schema
 */
export interface Google {
  /**
   * The type of authentication
   */
  type?: "google";
  /**
   * The id token of the user
   */
  idToken: string;
  /**
   * The audience of the user
   */
  audience: string;
}